Move crimegraphics files
