--- scrapers/CA/test.py
@@ -1,2 +1,0 @@
-from common import list_pdf_v3
-from common import list_pdf_v2
