+++ scrapers/data_portals/__init__.py
@@ -1,0 +1,8 @@
+#from .list_pdf_scrapers import list_pdf_v2
+#from .list_pdf_scrapers import list_pdf_v3
+#from .list_pdf_scrapers import single_pdf_scraper
+from .crimegraphics import crimegraphics_clery
+from .crimegraphics import crimegraphics_arrest
+from .crimegraphics import crimegraphics_bulletin
+#from .opendata import opendata_scraper
+#from .opendata import opendata_scraper2
