+++ scrapers/data_portals/crimegraphics/crimegraphics_arrest.py
